Add CoopAgreement model
Model to relate partner with a list of products available for their
sponsee partners. Add a code to identify the agreement.
Add permissions groups to manage the access right of this model.
We define two levels, one for create, read, write and unlink called
manager and one only with read access called user.